backend/requirements.txt
Django==4.1.10
dj_rest_auth==4.0.1
django-cors-headers
django_allauth==0.54.0
django_filer==2.2.5
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
django-bootstrap-sidebar
geoip2==4.7.0
mlflow==2.4.1
munch==3.0.0
numpy==1.25.0
opencv-python==4.7.0.72
Pillow==10.1.0
pix2tex==0.1.2
pytz==2023.3
PyYAML==6.0.1
Requests==2.31.0
# tensorboardX==2.6.2.2
torch==2.1.0
torchstat==0.0.7
tqdm==4.66.1

